wip moving stuff to helpers file
